refactor: Move and modify class according to the mvc pattern
